split: make separate trace and legacy utils
