+++ weave/trace/feedback.py
@@ -5,8 +5,7 @@ from typing import Any, Iterable, Iterator, Optional
 
 from rich.table import Table
 
-from weave.legacy.weave import util
-from weave.trace import rich_pydantic_util
+from weave.trace import rich_pydantic_util, util
 from weave.trace.client_context import weave_client as weave_client_context
 from weave.trace.refs import parse_uri
 from weave.trace.rich_container import AbstractRichContainer
